feat: move commandbar utils to dashboard/helper
