+++ app/javascript/dashboard/helper/commandbar/events.js
