+++ app/javascript/dashboard/helper/commandbar/icons.js
